[tests] Update tests to make native type vs normal type comparison explicit
We've modified our NUnitLite to special-case the native types, so that they
can easily be compared with the standard types, but that doesn't work anymore
when switching to the official NUnit[Lite], so change all asserts to
explicitly convert whenever needed.

+++ tests/monotouch-test/CoreFoundation/DispatchBlockTests.cs
@@ -100,11 +100,11 @@ namespace MonoTouchFixtures.CoreFoundation
 			using (var db = new DispatchBlock (callback)) {
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (0.1)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (0.1)));
 					Assert.AreNotEqual (0, rv, "Timed Out");
 
 					queue.DispatchAsync (db);
-					rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out 2");
 					Assert.IsTrue (called, "Called");
 				}
@@ -121,11 +121,11 @@ namespace MonoTouchFixtures.CoreFoundation
 			using (var db = new DispatchBlock (callback)) {
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
-					var rv = db.Wait (TimeSpan.FromSeconds (0.1));
+					var rv = (int) db.Wait (TimeSpan.FromSeconds (0.1));
 					Assert.AreNotEqual (0, rv, "Timed Out");
 
 					queue.DispatchAsync (db);
-					rv = db.Wait (TimeSpan.FromSeconds (5));
+					rv = (int) db.Wait (TimeSpan.FromSeconds (5));
 					Assert.AreEqual (0, rv, "Timed Out 2");
 					Assert.IsTrue (called, "Called");
 				}
@@ -138,10 +138,10 @@ namespace MonoTouchFixtures.CoreFoundation
 			var called = false;
 			var callback = new Action (() => called = true);
 			using (var db = new DispatchBlock (callback)) {
-				Assert.AreEqual (0, db.TestCancel (), "TestCancel 1");
+				Assert.AreEqual ((nint) 0, db.TestCancel (), "TestCancel 1");
 				Assert.IsFalse (db.Cancelled, "Cancelled 1");
 				db.Cancel ();
-				Assert.AreNotEqual (0, db.TestCancel (), "TestCancel 2");
+				Assert.AreNotEqual ((nint) 0, db.TestCancel (), "TestCancel 2");
 				Assert.IsTrue (db.Cancelled, "Cancelled 2");
 			}
 		}
@@ -170,7 +170,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out A");
 					Assert.IsTrue (called, "Called A");
 				}
@@ -182,7 +182,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -194,7 +194,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -207,7 +207,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -219,7 +219,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out Background 8" + flags);
 					Assert.IsTrue (called, "Called Background 8" + flags);
 				}
@@ -250,7 +250,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out A");
 					Assert.IsTrue (called, "Called A");
 				}
@@ -262,7 +262,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -274,7 +274,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -287,7 +287,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out " + flags);
 					Assert.IsTrue (called, "Called " + flags);
 				}
@@ -299,7 +299,7 @@ namespace MonoTouchFixtures.CoreFoundation
 				using (var queue = new DispatchQueue ("Background")) {
 					queue.Activate ();
 					queue.DispatchAsync (db);
-					var rv = db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+					var rv = (int) db.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 					Assert.AreEqual (0, rv, "Timed Out Background 8" + flags);
 					Assert.IsTrue (called, "Called Background 8" + flags);
 				}
@@ -312,7 +312,7 @@ namespace MonoTouchFixtures.CoreFoundation
 					using (var queue = new DispatchQueue ("Background")) {
 						queue.Activate ();
 						queue.DispatchAsync (db2);
-						var rv = db2.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
+						var rv = (int) db2.Wait (new DispatchTime (DispatchTime.Now, TimeSpan.FromSeconds (5)));
 						Assert.AreEqual (0, rv, "Timed Out Background DB" + flags);
 						Assert.IsTrue (called, "Called Background DB" + flags);
 					}
